+++ pages/api/elasticSearch.js
@@ -1,4 +1,4 @@
 const index_name = "cassandra_job_posts";
-const esEndpoint = `http://${process.env.API_ENDPOINT}/api/elasticsearch/${index_name}`;
+const esEndpoint = `http://167.172.142.105:5000/api/elasticsearch/${index_name}`;
 
 export default esEndpoint;
